adjusting CreateOnlyAttributePlanModifier to support all attribute types
