+++ internal/service/project/resource_project_schema.go
@@ -53,7 +53,7 @@ func ResourceSchema(ctx context.Context) schema.Schema {
 			"project_owner_id": schema.StringAttribute{
 				Optional: true,
 				PlanModifiers: []planmodifier.String{
-					customplanmodifier.CreateOnlyStringPlanModifier(),
+					customplanmodifier.CreateOnlyAttributePlanModifier(),
 				},
 			},
 			"with_default_alerts_settings": schema.BoolAttribute{
